online_shop/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
az-iranian-bank-gateways==1.6.15
beautifultable==0.8.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.2
constantly==15.1.0
coverage==6.3.1
cryptography==36.0.1
daphne==3.0.2
Django==3.2.12
django-admin-volt==0.0.5
django-crispy-forms==1.14.0
django-location-field==2.1.0
django-ranged-response==0.2.0
django-rosetta==0.9.8
django-simple-captcha==0.5.17
djangorestframework==3.13.1
fandogh-cli==1.26.10
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isodate==0.6.1
kavenegar==1.1.2
lxml==4.8.0
msgpack==1.0.3
outcome==1.1.0
Pillow==9.0.1
platformdirs==2.5.1
polib==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pyOpenSSL==22.0.0
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
retrying==1.3.3
selenium==4.1.0
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
tzlocal==2.1
urllib3==1.26.8
websocket-client==1.3.2
wsproto==1.0.0
zeep==4.1.0
zope.interface==5.4.0
